requirements/local.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
Django==3.2.3
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
idna==3.2
mercadopago==2.0.7
Pillow==8.2.0
pycparser==2.20
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
Unipath==1.1
urllib3==1.26.6
